// src/services/models/MedicalVisitModel.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type MedicalVisitModel = {
    id?: string | null;
    infantId?: string;
    complaint?: string | null;
    date?: string;
    clinicalTeamId?: string;
    diagnosis?: string | null;
    physicianExamination?: string | null;
    medication?: string | null;
    heartRate?: number;
    respiratoryRate?: number;
    bloodPressure?: number;
    temperature?: number;
    oxygenSaturation?: number;
};